+++ src/stones/LightPassengerStone.cc
@@ -42,17 +42,17 @@ namespace enigma {
             setState(1 - state);
             return Value();
         } else if (m.message == "_model_reanimated") {
-            objFlags &= ~OBJBIT_SKATEDIR;  // NODIR
-            if (state != BREAK) {     // continue break animation
-                init_model();         // adjust possible changed model in other cases
+            objFlags &= ~OBJBIT_SKATEDIR; // NODIR
+            if (state != BREAK) {         // continue break animation
+                init_model();             // adjust possible changed model in other cases
             }
-            return Value();            
+            return Value();
         } else if (m.message == "_init") {
             if (objFlags & OBJBIT_LIGHTNEWDIRS)
                 lightDirChanged(NODIRBIT, (DirectionBits)(objFlags & OBJBIT_LIGHTNEWDIRS));
-            return Value();            
+            return Value();
         } else if (m.message == "_glasses") {
-            if ((((int)(m.value) & 16) != 0) != ((objFlags & OBJBIT_VISIBLE) != 0)) {
+            if (((m.value.toInt() & 16) != 0) != ((objFlags & OBJBIT_VISIBLE) != 0)) {
                 objFlags ^= OBJBIT_VISIBLE; // toggle visibility bit
                 init_model();
             }
@@ -241,10 +241,10 @@ namespace enigma {
             the force resulting from floor->add_force. "baseinterval"
             is 50 ms or the interval given in "interval".
         */
-        double base = getAttr("interval");
+        double base = getAttr("interval").toDouble();
         if (Floor *floor = GetFloor(get_pos())) {
             if (Value f = getAttr("friction"))
-                base *= 1.0 + (double)f * floor->get_friction();
+                base *= 1.0 + f.toDouble() * floor->get_friction();
             if (Value g = getAttr("gradient")) {
                 Direction skateDir = (Direction)((int)((objFlags & OBJBIT_SKATEDIR) >> 24) - 1);
                 if (skateDir != NODIR) {
@@ -253,7 +253,7 @@ namespace enigma {
                     floor->add_force(0, vec);
                     quot = skateDir == NORTH ? -vec[1] : skateDir == SOUTH ? vec[1] :
                         skateDir == EAST ? vec[0] : skateDir == WEST ? -vec[0] : 0;
-                    base /= std::max(1.0 + (double)g * quot, 0.01);                    
+                    base /= std::max(1.0 + g.toDouble() * quot, 0.01);
                 }
             }
         }
